Book: Make test chapter code work as tests
