--- src/tutorial/testing-working.rs
@@ -1,16 +1,0 @@
-#![allow(unused)]
-
-fn find_matches<W: std::io::Write>(content: &str, pattern: &str, writer: &mut W) {
-    for line in content.lines() {
-        if line.contains(pattern) {
-            writeln!(writer, "{}", line);
-        }
-    }
-}
-
-#[test]
-fn find_a_match() {
-    let mut result = Vec::new();
-    find_matches("lorem ipsum\ndolor sit amet", "lorem", &mut result);
-    assert_eq!(result, b"lorem ipsum\n");
-}
