+++ src/tutorial/testing/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "CLAiR-testing"
+version = "0.1.0"
+authors = ["Pascal Hertleif <killercup@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+structopt = "0.2.10"
+failure = "0.1.3"
+exitfailure = "0.5.1"
+
+[dev-dependencies]
+assert_cmd = "0.10"
+predicates = "1"
+tempfile = "3"
