+++ src/tutorial/testing/tests/cli.rs
@@ -1,5 +1,3 @@
-#![allow(unused)]
-
 use std::process::Command;  // Run programs
 use assert_cmd::prelude::*; // Add methods on commands
 use predicates::prelude::*; // Used for writing assertions
@@ -12,6 +10,24 @@ fn file_doesnt_exist() -> Result<(), Box<std::error::Error>> {
     cmd.assert()
         .failure()
         .stderr(predicate::str::contains("No such file or directory"));
+
+    Ok(())
+}
+
+use tempfile::NamedTempFile;
+use std::io::{self, Write};
+
+#[test]
+fn find_content_in_file() -> Result<(), Box<std::error::Error>> {
+    let mut file = NamedTempFile::new()?;
+    writeln!(file, "A test\nActual content\nMore content\nAnother test")?;
+
+    let mut cmd = Command::main_binary()?;
+    cmd.arg("test")
+        .arg(file.path());
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains("test\nAnother test"));
 
     Ok(())
 }
